CLOUD-155: Add GitLab for Enterprise to Dashboard GitLab Integration docs
This change adds minimal edits to the **Dashboard -> GitLab Integration** page to briefly explain pitfalls & additional steps specific to the new GitLab for Enterprise integration.

Please note that this change is *not* ready to be merged as there are outstanding issues to be resolved around the branding of this feature - in GitLab nomenclature "Enterprise" refers to a functionally different edition of their core software, **not** to whether it can be self-hosted.

Furthermore, the Dashboard's integration configuration areas are currently undergoing redesign, so consider this a precursor to a more thorough rewrite of the integration docs that will address their existing shortcomings.